+++ src/main/java/de/binary101/core/data/item/Shield.java
@@ -6,10 +6,10 @@ public class Shield extends Item {
 
 	@Getter int blockChance;
 	
-	public Shield(Integer[] itemData, int backpackIndex) {
+	public Shield(Long[] itemData, int backpackIndex) {
 		super(itemData, backpackIndex);
 		
-		this.blockChance = itemData[2];
+		this.blockChance = itemData[2].intValue();
 	}
 	
 	@Override
